PrismDB_Lite/requirements.txt
# Core Framework
flask==2.3.3
flask-cors==4.0.0
flask-sse==1.0.0
agno==0.1.3
pyyaml==6.0.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# AI APIs
openai==1.3.7
google-generativeai==0.3.1

# Security
python-dotenv==1.0.0
pyjwt==2.8.0

# Performance
redis==5.0.1
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0 